allow extrn symbols to be defined by the unit itself
Makes it possible to put all the .extrn declarations
in a shared header file and include that header file
in both the implementation of a unit and by users of
that unit.

I just got really sick of having to both include the
header _and_ explicitly declare the .extrns every time.